Updated testing files for cli
